refactor: restructure server code
