+++ package.json
@@ -14,6 +14,7 @@
   },
   "main": "server.js",
   "dependencies": {
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "passport": "^0.4.1",
@@ -40,8 +41,8 @@
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
     "typescript": "~4.1.5",
-    "vue-jest": "^5.0.0-0",
     "vue": "^3.0.0",
+    "vue-jest": "^5.0.0-0",
     "vuex": "^4.0.0-0"
   },
   "_id": "connecting-senses@0.0.1",
